version 4.0.0-nitroball community compilada

+++ sources/modules/control-panel/src/main/java/com/minsait/onesait/platform/controlpanel/controller/multitenant/TenantDTO.java
@@ -29,5 +29,5 @@ public class TenantDTO {
 
 	private String name;
 	private List<String> vertical;
-	private int users;
+	private long users;
 }
